Also upgrade Kotlin to 1.4.0

+++ alchemist/alchemist-euclidean-geometry/src/test/kotlin/it/unibo/alchemist/test/TestRandomTarget.kt
@@ -78,7 +78,7 @@ private class DummyRandomGenerator : RandomGenerator {
     }
 
     override fun nextBytes(bytes: ByteArray) = bytes.forEachIndexed { idx, _ ->
-        bytes[idx] = value.toByte()
+        bytes[idx] = value.toInt().toByte()
     }
 
     override fun nextInt() = value.toInt()
